requirements.txt
asgiref==3.2.10
autopep8==1.4.4
Django==3.0.9
django-appconf==1.0.3
django-cors-headers==3.4.0
django-imagekit==4.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-nested-routers==0.91
mysqlclient==2.0.1
pilkit==2.0
Pillow==6.2.1
psycopg2-binary==2.8.4
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0